requirements.txt
pytorch-lightning>=1.5.5
sacred>=0.8.2
datasets>=1.16.1
torchvision>=0.11.1
pycocotools>=2.0.3
gensim>=4.1.2
transformers>=4.13.0
sklearn>=1.0.1
terminaltables>=3.1.10
openpyxl>=3.0.9
fasttext>=0.9.2
jupyterlab>=3.2.8
nltk>=3.6.7
torchinfo>=1.6.3
wandb>=0.12.11
seaborn>=0.11.2
